+++ bin/test.sh
@@ -10,9 +10,9 @@ echo "##################################################" >> ../onlysnarf/logs/t
 echo "################## Start #########################" >> ../onlysnarf/logs/test.txt 2>&1
 echo "##################################################" >> ../onlysnarf/logs/test.txt 2>&1
 
-sudo onlysnarfpy -debug -verbose -verboser -type gallery -debug-delay -show-window -skip-reduce
+# sudo onlysnarfpy -debug -verbose -verboser -type gallery -debug-delay -show-window -skip-reduce
 
-# sudo onlysnarf -debug -verbose -show-window -debug-delay -prefer-local
+sudo onlysnarf -debug -verbose -show-window -debug-delay -prefer-local
 # sudo onlysnarfpy -verbose -type gallery -notkeyword feet -show-window -debug -debug-delay
 # sudo onlysnarf-config
 # sudo onlysnarfpy -debug -verbose -action test